fix: providers on a module level
BREAKING CHANGE:

Component providers should be provided via `componentProviders`.
